requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-chess==1.10.0
werkzeug==3.0.1 